feat: showMore modify
